Add user list page

+++ application/models/central_admin/users_model.php
@@ -20,6 +20,60 @@ class Users_model extends CI_Model {
 	}
 
 	/**
+	 * Get existing users
+	 *
+	 * @access	public
+	 * @param	int		Page number for the list
+	 * @return	array	List of users
+	 */
+	public function get_users($page)
+	{
+		$per_page = $this->config->item('per_page');
+		$offset = $per_page * ($page - 1);
+		$filter = $this->input->post('user_filter');
+		
+		if ( ! empty($filter))
+		{
+			$this->db_c->like('user_name', $filter);
+		}
+
+		$query = $this->db_c->limit($per_page, $offset)->order_by('user_name')->get('users');
+		return $query->result();
+	}
+
+	/**
+	 * Return the count of users from the DB
+	 *
+	 * @access	public
+	 * @return	int		having the user count
+	 */
+	public function count_users()
+	{
+		return $this->db_c->count_all_results('users');
+	}
+
+	/**
+	 * Checks if a user is the founder
+	 *
+	 * @access	public
+	 * @param	int		User ID
+	 * @return	bool	true if user is founder
+	 */
+	public function check_founder($user_id)
+	{
+		$query = $this->db_c->get_where('users', array('user_id' => $user_id));
+
+		if ($query->num_rows() == 1)
+		{
+			return $query->row()->user_founder == 1;
+		}
+		else
+		{
+			return FALSE;
+		}
+	}
+
+	/**
 	 * Get user details against a specific user ID
 	 *
 	 * @access	public
@@ -82,8 +136,7 @@ class Users_model extends CI_Model {
 			$data['user_password'] = $password_hash;
 		}
 
-		$this->db_c->where('user_id', $user_id);
-		return $this->db_c->update('users', $data);
+		return $this->db_c->where('user_id', $user_id)->update('users', $data);
 	}
 
 	/**
@@ -94,8 +147,6 @@ class Users_model extends CI_Model {
 	 */
 	public function delete_user($user_id)
 	{
-		$this->db_c->where('user_id', $user_id);
-
-		return $this->db_c->delete('users');
+		return $this->db_c->where('user_id', $user_id)->delete('users');
 	}
 }
